requirements.txt
streamlit>=1.26.0
openai-whisper
moviepy==1.0.3
opencv-python-headless
pillow
numpy
srt
